[change] move all newsletter validation tests to one module
