--- tests/site/newsletter/conftest.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-@pytest.fixture(scope="package")
-def dummy_email() -> str:
-    return "dummy@dummydomain.com"
